Rename FhirTaskUtil to FhirResourceUtil
