+++ android/engine/src/test/java/org/smartregister/fhircore/engine/task/FhirTaskUtilTest.kt
@@ -46,6 +46,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
+import org.smartregister.fhircore.engine.configuration.ConfigurationRegistry
 import org.smartregister.fhircore.engine.data.local.DefaultRepository
 import org.smartregister.fhircore.engine.robolectric.RobolectricTest
 import org.smartregister.fhircore.engine.util.extension.isIn
@@ -57,24 +58,24 @@ import org.smartregister.fhircore.engine.util.extension.today
 class FhirTaskUtilTest : RobolectricTest() {
 
   @get:Rule(order = 0) val hiltAndroidRule = HiltAndroidRule(this)
-  private lateinit var fhirTaskUtil: FhirTaskUtil
+  private lateinit var fhirTaskUtil: FhirResourceUtil
   private lateinit var fhirEngine: FhirEngine
   private lateinit var defaultRepository: DefaultRepository
-  private lateinit var fhirResourceClosureUtil: FhirResourceClosureUtil
+  private lateinit var configurationRegistry: ConfigurationRegistry
 
   @Before
   fun setup() {
     hiltAndroidRule.inject()
     fhirEngine = spyk(FhirEngineProvider.getInstance(ApplicationProvider.getApplicationContext()))
     defaultRepository = mockk()
-    fhirResourceClosureUtil = mockk()
+    configurationRegistry = mockk()
     every { defaultRepository.fhirEngine } returns fhirEngine
     fhirTaskUtil =
       spyk(
-        FhirTaskUtil(
+        FhirResourceUtil(
           ApplicationProvider.getApplicationContext(),
           defaultRepository,
-          fhirResourceClosureUtil,
+          configurationRegistry,
         ),
       )
   }
